Fixing code duplication issue on models patient and responsible
